moving evaluate methods from EvalMixin subclasses to  eval.expression
