+++ mathics/core/symbols.py
@@ -1,7 +1,6 @@
 # cython: language_level=3
 # -*- coding: utf-8 -*-
 
-import time
 from typing import Any, FrozenSet, List, Optional
 
 from mathics.core.element import (
@@ -472,23 +471,9 @@ class Symbol(Atom, NumericOperators, EvalMixin):
         Evaluates the symbol by applying the rules (ownvalues) in its definition,
         recursively.
         """
-        if evaluation.definitions.trace_evaluation:
-            if evaluation.definitions.timing_trace_evaluation:
-                evaluation.print_out(time.time() - evaluation.start_time)
-            evaluation.print_out(
-                "  " * evaluation.recursion_depth + "  Evaluating: %s" % self
-            )
+        from mathics.eval.expression import evaluate_symbol
 
-        rules = evaluation.definitions.get_ownvalues(self.name)
-        for rule in rules:
-            result = rule.apply(self, evaluation, fully=True)
-            if result is not None and not result.sameQ(self):
-                if evaluation.definitions.trace_evaluation:
-                    evaluation.print_out(
-                        "  " * evaluation.recursion_depth + "  -> %s" % result
-                    )
-                return result.evaluate(evaluation)
-        return self
+        return evaluate_symbol(self, evaluation)
 
     def get_head(self) -> "Symbol":
         return Symbol("Symbol")
